shiv-agent/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
google-generativeai==0.5.4
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.3
langdetect==1.0.9
